+++ CHANGES.md
@@ -4,6 +4,7 @@
 ## Unreleased
 - MongoDB: Unlock processing multiple collections, either from server database,
   or from filesystem directory
+- MongoDB: Unlock processing JSON files from HTTP resource, using `https+bson://`
 
 ## 2024/09/10 v0.0.22
 - MongoDB: Rename columns with leading underscores to use double leading underscores